Crate eq_session_manager

Equilibrium Session Manager Pallet

Equilibrium's Session Manager Pallet is a Substrate module that manages validation of Equilibrium substrate

Structs

GenesisConfig
Genesis config for the module, allow to build genesis storage.

IsChanged
Pallet storage - flag showing that active validators list changed during a session

Module

Validators
Pallet storage - list of all active validators

Enums

Call
Dispatchable calls.

Error
Validator registration errors

RawEvent
Events for this module.

Traits

Trait
Substrate pallet configuration trait

Type Definitions

Event
RawEvent specialized for the configuration Trait